rename to index.js for components and containers
