+++ src/client.js
@@ -59,7 +59,7 @@ if (process.env.NODE_ENV !== 'production') {
 }
 
 if (__DEVTOOLS__ && !window.devToolsExtension) {
-  const DevTools = require('./containers/DevTools/DevTools');
+  const DevTools = require('./containers/DevTools');
   ReactDOM.render(
     <Provider store={store} key="provider">
       <div>
