+++ src/components/index.js
@@ -5,9 +5,9 @@
  *
  */
 
-export CounterButton from './CounterButton/CounterButton';
-export GithubButton from './GithubButton/GithubButton';
-export InfoBar from './InfoBar/InfoBar';
-export MiniInfoBar from './MiniInfoBar/MiniInfoBar';
-export SurveyForm from './SurveyForm/SurveyForm';
-export WidgetForm from './WidgetForm/WidgetForm';
+export CounterButton from './CounterButton';
+export GithubButton from './GithubButton';
+export InfoBar from './InfoBar';
+export MiniInfoBar from './MiniInfoBar';
+export SurveyForm from './SurveyForm';
+export WidgetForm from './WidgetForm';
