+++ src/containers/index.js
@@ -1,9 +1,9 @@
-export App from './App/App';
-export Chat from './Chat/Chat';
-export Home from './Home/Home';
-export Widgets from './Widgets/Widgets';
-export About from './About/About';
-export Login from './Login/Login';
-export LoginSuccess from './LoginSuccess/LoginSuccess';
-export Survey from './Survey/Survey';
-export NotFound from './NotFound/NotFound';
+export App from './App';
+export Chat from './Chat';
+export Home from './Home';
+export Widgets from './Widgets';
+export About from './About';
+export Login from './Login';
+export LoginSuccess from './LoginSuccess';
+export Survey from './Survey';
+export NotFound from './NotFound';
